Add files via upload
This project “Instagram User Analytics” tracks how users engage and interact with our digital product i.e. Instagram (in this case).
The project attempts to derive precise and intelligent business insights for marketing, product & development teams.
The insights gained are then can used by teams across the business to launch a new marketing campaign, decide on features to build, maintain and improvise Instagram to track the success of the platform by measuring user engagement and improve the experience altogether while helping the business grow.
In this I have imagined working with the product team of Instagram and the product manager who has asked me to provide insights on the questions asked by the management team.