Add files via upload
In this project, I learned how to apply advanced SQL concepts like Windows Functions, etc. I understood how the real-world industry works. It helped me in mastering my SQL concepts. I learned how to ask the right questions given the circumstances. From the given data and questions, which columns to consider and how to find the valuable insights which help the business to grow. I learned how the company finds different areas related to the company to improve it further. I got to know about investigating metrics spikes (why there is a boom and why there is a dip).