Add files via upload
We have here a dataset for IMDB Movies. I have framed the problem. For this task, I have defined a problem I want to shed light on.
I did this by asking 'What?' i.e. What is the problem? Eg. What do I see happening? What is my hypothesis for the cause of the problem? What is the impact of the problem on stakeholders? What is the impact of the problem not being solved?
Answering these questions helped define the problem I was trying to solve and will be allowed me to find the right data to solve it.
Now next step I cleaned the data as necessary, and use your Data Analysis skills to explore the data set and derive insights.
Using Advanced Excel and Statistical concepts we have found movies with the highest profit, top movies as per IMDb rating, top directors, most popular genres, top foreign language films, and more.